Testing

+++ app/Http/Controllers/UserController.php
@@ -2,10 +2,12 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Tender;
 use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Hash;
+use Illuminate\Support\Facades\Redirect;
 
 class UserController extends Controller
 {
@@ -18,10 +20,32 @@ class UserController extends Controller
         $credentials = $request->only('email', 'password');
 
         if (Auth::attempt($credentials)) {
-            return redirect()->intended('/home');
+
+            $userCount = User::count() ?? 0;
+
+            // Get the counts for different categories in the Tender model
+            $localCount = Tender::where('Category', 1)->count() ?? 0;
+            $foreignCount = Tender::where('Category', 2)->count() ?? 0;
+            $otherCount = Tender::where('Category', 3)->count() ?? 0;
+
+            // Pass the counts to the view
+            return view('admin.home')->with(compact('userCount', 'localCount', 'foreignCount', 'otherCount'));
         }
 
-        return redirect()->back()->withInput()->withErrors(['error' => 'Invalid credentials. Check your user name and password']);
+        return redirect()->back()->withInput()->withErrors(['error' => 'Invalid credentials. Check your user E-Mail and Password']);
+    }
+    //pass dashboard count values
+    public function dashboard()
+    {
+        $userCount = User::count() ?? 0;
+
+        // Get the counts for different categories in the Tender model
+        $localCount = Tender::where('Category', 1)->count() ?? 0;
+        $foreignCount = Tender::where('Category', 2)->count() ?? 0;
+        $otherCount = Tender::where('Category', 3)->count() ?? 0;
+
+        // Pass the counts to the view
+        return view('admin.home')->with(compact('userCount', 'localCount', 'foreignCount', 'otherCount'));
     }
 
     public function logout()
@@ -32,8 +56,19 @@ class UserController extends Controller
 
     public function userdelete($id)
     {
-        $record = User::orderBy('id', 'asc')->find($id);
-        $record->delete();
+        // Find the user record by ID
+        $record = User::find($id);
+
+        // Check if the email is the one you want to prevent from deletion
+        if ($record && $record->email == 'mainadmin99@slpa.lk') {
+            return redirect()->back()->with('error', 'Cannot delete Main Admin');
+        }
+
+        // If the email is not the one to be prevented, delete the user record
+        if ($record) {
+            $record->delete();
+        }
+
         return redirect()->back();
     }
 
@@ -70,5 +105,6 @@ class UserController extends Controller
     {
         return view('admin.adduser');
     }
+
 
 }
